Convert discrete inventory counts to int; finalize README
localInventory (StoreFront) and loadedQuantity (Truck) are now int.
Demand and load quantities are rounded at the source with Math.round()
and cast to int, ensuring all product counts are integers throughout
the simulation. README updated to match the final implemented design.

+++ SupplyChain/_alp/Agents/Truck/Code/Functions.java
@@ -1,6 +1,6 @@
 double tryLoad()
 {/*ALCODESTART::1774650002000*/
-double quantity = Math.round(Math.min(main.maxInventory - assignedStore.localInventory, main.products));
+int quantity = (int) Math.round(Math.min(main.maxInventory - assignedStore.localInventory, main.products));
 main.products -= quantity;
 loadedQuantity = quantity;
 /*ALCODEEND*/}
